Beta klar

+++ src/main/java/se/mau/webbserver/entity/attendance/AttendanceRepository.java
@@ -2,5 +2,5 @@ package se.mau.webbserver.entity.attendance;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
+public interface AttendanceRepository extends JpaRepository<Attendance, Integer> {
 }
