+++ src/main/java/se/mau/webbserver/entity/cost_center/CostCenterRepository.java
@@ -2,5 +2,5 @@ package se.mau.webbserver.entity.cost_center;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CostCenterRepository extends JpaRepository<CostCenter, Long> {
+public interface CostCenterRepository extends JpaRepository<CostCenter, Integer> {
 }
